+++ server/channels/store/sqlstore/migrate.go
@@ -12,6 +12,8 @@ import (
 	"strconv"
 	"sync"
 
+	"github.com/mattermost/mattermost/server/public/utils"
+
 	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/mattermost/mattermost/server/public/shared/mlog"
 	"github.com/mattermost/mattermost/server/v8/channels/db"
@@ -119,12 +121,12 @@ func (ss *SqlStore) initMorph(dryRun bool) (*morph.Morph, error) {
 	var driver drivers.Driver
 	switch ss.DriverName() {
 	case model.DatabaseDriverMysql:
-		dataSource, rErr := ResetReadTimeout(*ss.settings.DataSource)
+		dataSource, rErr := utils.ResetReadTimeout(*ss.settings.DataSource)
 		if rErr != nil {
 			mlog.Fatal("Failed to reset read timeout from datasource.", mlog.Err(rErr), mlog.String("src", *ss.settings.DataSource))
 			return nil, rErr
 		}
-		dataSource, err = AppendMultipleStatementsFlag(dataSource)
+		dataSource, err = utils.AppendMultipleStatementsFlag(dataSource)
 		if err != nil {
 			return nil, err
 		}
