+++ server/channels/store/sqlstore/store.go
@@ -15,6 +15,8 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/mattermost/mattermost/server/public/utils"
+
 	sq "github.com/mattermost/squirrel"
 
 	"github.com/go-sql-driver/mysql"
@@ -314,7 +316,7 @@ func (ss *SqlStore) initConnection() error {
 		// covers that already. Ideally we'd like to do this only for the upgrade
 		// step. To be reviewed in MM-35789.
 		var err error
-		dataSource, err = ResetReadTimeout(dataSource)
+		dataSource, err = utils.ResetReadTimeout(dataSource)
 		if err != nil {
 			return errors.Wrap(err, "failed to reset read timeout from datasource")
 		}
